Solved linter issues
