+++ source/Modules/index.js
@@ -1,2 +1,2 @@
 export * from './storageManager';
-export * from './Controller';
+export * from './Controller/connection';
